backend/masters/serializers.py
from rest_framework import serializers
from .models import EmployeeType

class EmployeeTypeSerializer(serializers.ModelSerializer):
    class Meta:
        model = EmployeeType
        fields = '__all__'
